Make ProjectionSelectionPanel a manual layout
Remove the Netbeans form definition file. The form and the Java source
were not syncable by Netbeans anymore because the Java source has been
modified after it had been generated from the .form file. Without the
ability to make changes to the layout through the graphical form editor
the Netbeans tools don't help much when working on the form.

Work on the generated source to make it more readable and easier to edit
by hand:

* Use imports instead of fully qualified references
* Remove comments created by the form generator
* Move fields to the top of the class